Made all currencies into doubles to make smaller values possible

+++ src/main/java/com/petometry/petservice/service/model/currency/CurrencyBalance.java
@@ -9,5 +9,5 @@ public class CurrencyBalance {
 
     private CurrencyType currency;
 
-    private Long balance;
+    private Double balance;
 }
